Move assign_targets_from_sb40 to landfire module
