+++ duet_tools/inputs.py
@@ -6,11 +6,9 @@ from __future__ import annotations
 
 # Core Imports
 from pathlib import Path
-import warnings
 from random import randint
 
 # External Imports
-import numpy as np
 
 # Internal Imports
